# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pingit API",
        "description": "Pingit is a full-stack uptime monitoring system. Monitor your URLs, get notified when they go down, and track performance over time.",
        "contact": {
            "email": "support@pingit.live"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://pingit.live",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/monitors": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "List all monitors for the authenticated user",
                "operationId": "b238bf9af384f617bb4c2bbc47fd076a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "up",
                                "down",
                                "degraded",
                                "paused"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "created_at",
                                "last_checked_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by tag name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of monitors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Register a new URL to monitor",
                "operationId": "9e798d4fe6fd77ea752fde01ecbcc27b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "My Website"
                                    },
                                    "check_interval": {
                                        "type": "integer",
                                        "maximum": 60,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "threshold": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 3
                                    },
                                    "response_time_threshold_ms": {
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "http_method": {
                                        "type": "string",
                                        "enum": [
                                            "GET",
                                            "HEAD"
                                        ],
                                        "example": "GET"
                                    },
                                    "follow_redirects": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Monitor created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{id}": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get a single monitor",
                "operationId": "2c53059cb1dfc4eea2e54eb0c0884ab8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor details"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Update a monitor",
                "operationId": "a4b03518b1e1e0479d86742725d2b461",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "check_interval": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "threshold": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Monitor updated successfully"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Delete a monitor",
                "operationId": "d65b15ae84bf2951088598c99cd1bd20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor deleted successfully"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{id}/pause": {
            "post": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Pause a monitor",
                "operationId": "801a17816eb53c4458b4616edc9432a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor paused"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{id}/resume": {
            "post": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Resume a paused monitor",
                "operationId": "177d8570f57778f8935fef8ecbf187b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monitor resumed"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{id}/history": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get paginated check history for a monitor",
                "operationId": "2a310ce550e024727d0e29fa13f3576a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check history"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{id}/history/export": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Export check history as CSV",
                "operationId": "d1d6b181275b0d1513746cd7abbdc428",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file download"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{id}/incidents": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get incident history for a monitor",
                "operationId": "e67de35bfe6778e006de040b10fe5174",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incident history"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/monitors/{id}/response-times": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get response time trends for a monitor",
                "operationId": "070685c7cdaef3f026f67fbbf8207cf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Time period",
                        "schema": {
                            "type": "string",
                            "default": "7d",
                            "enum": [
                                "24h",
                                "7d",
                                "30d"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response time trends"
                    },
                    "404": {
                        "description": "Monitor not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}